Rename Filter WIP
def get_parameter_name_by_id added.
The dictionary of the parameter is not need apparently.
With this one, we can get the Workset parameter name from the parameter ID.
The workset is not property in the property set for the types or non types. It is a parameter ("property") of the element.